chore: move AnimatorLayerMap to Utils
