+++ Editor/Processors/TraceAndOptimize/AnimatorParser.cs
@@ -477,43 +477,6 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
             return controller;
         }
 
-        private class AnimatorLayerMap<T>
-        {
-            private T[] _values = new T[(int)(VRCAvatarDescriptor.AnimLayerType.IKPose + 1)];
-
-            public static bool IsValid(VRCAvatarDescriptor.AnimLayerType type)
-            {
-                switch (type)
-                {
-                    case VRCAvatarDescriptor.AnimLayerType.Base:
-                    case VRCAvatarDescriptor.AnimLayerType.Additive:
-                    case VRCAvatarDescriptor.AnimLayerType.Gesture:
-                    case VRCAvatarDescriptor.AnimLayerType.Action:
-                    case VRCAvatarDescriptor.AnimLayerType.FX:
-                    case VRCAvatarDescriptor.AnimLayerType.Sitting:
-                    case VRCAvatarDescriptor.AnimLayerType.TPose:
-                    case VRCAvatarDescriptor.AnimLayerType.IKPose:
-                        return true;
-                    case VRCAvatarDescriptor.AnimLayerType.Deprecated0:
-                    default:
-                        return false;
-                }
-            }
-
-            public ref T this[VRCAvatarDescriptor.AnimLayerType type]
-            {
-                get
-                {
-                    if (!IsValid(type))
-                        throw new ArgumentOutOfRangeException(nameof(type), type, null);
-
-                    return ref _values[(int)type];
-                }
-            }
-        }
-
-        private static readonly int LayersCount = (int)(VRCAvatarDescriptor.AnimLayerType.IKPose + 1);
-
         private static readonly AnimatorLayerMap<CachedGuidLoader<AnimatorController>> DefaultLayers =
             new AnimatorLayerMap<CachedGuidLoader<AnimatorController>>
             {
